feat(ios): keyboard light/dark, shared fullwidth input and candidate font chain (#584)
* feat(ios): keyboard and panel light/dark from the shared themes

The skin page gains a 「明暗」 section for the shared `screen_keyboard_theme`, plus the handwriting, emoji and voice panel themes. The keyboard applies them as its interface style, so every dynamic skin colour draws in the chosen form whatever the host app looks like. `follow` falls back to the global `theme`, then to the host; a panel left on `follow` follows the keyboard. iPad lists the panel pickers inline; iPhone folds them into a disclosure group.

* feat(ios): route fullwidth input through the shared preference and runtime

The iOS keyboard kept its 全角 switch in the App Group key `keyboard.input.fullWidth` and never called `msime_client_set_character_width`. The shared `character_width` setting did nothing on iOS, and everything the runtime committed stayed halfwidth.

It now works like the other five hosts. The shared `character_width` is the width a keyboard starts in; the keyboard tells the runtime, and the session restores the width whenever it is rebuilt (every time the keyboard reappears). The 「更多」 card is a switch for the running keyboard only, and a reload replaces it only when `character_width` itself changed. The App's 标点 page gains the 全角输入 toggle, and on launch the App folds the old App Group switch into the document once. The `character_width` host capability is now on for iOS as well.

* feat(ios): candidate font family chain from the shared document

The candidate strip and panel now draw in the shared `candidate_english_font`, `candidate_font_family` and `candidate_fallback_fonts`, chained the way the desktop candidate window does it. The Latin face leads, and the family and its fallbacks follow in a `UIFontDescriptor` cascade list. Faces the device lacks are skipped, so a document synced from Windows, or the untouched default Noto Sans SC, keeps the system font. The chain scales with the candidate size and Dynamic Type.

The App's candidate page gains 中文字体 and 英文字体 pickers. Each lists this device's families, drawn in themselves. A synced name the device lacks stays listed and is marked as not installed. The `candidate_font_controls` and `candidate_english_font` capabilities are now on for iOS.

* fix(build): adapt to the sha2 0.11 and hmac 0.13 APIs

The rustcrypto bump (#542) left develop uncompilable: sha2 0.11 digests are hybrid arrays that no longer implement LowerHex, and hmac 0.13 moved new_from_slice onto the KeyInit trait. client-core now encodes digests with its existing hex dependency; engine-bridge, host-api and the Android account plugin, which have no hex dependency, use a local lower-hex helper instead of gaining one. The rusqlite 0.40 bump (#538) also dropped ToSql for usize, so the host-api fixtures bind their row index as i64.

+++ crates/host-api/src/tests.rs
@@ -3636,17 +3636,17 @@ fn emoji_catalog_cursor_advances_over_invalid_rows_and_preserves_duplicates() {
     {
         db.execute(
             "INSERT INTO emoji VALUES (?1,'fixture','match','',?2)",
-            rusqlite::params![text, index],
+            rusqlite::params![text, index as i64],
         )
         .unwrap();
         db.execute(
             "INSERT INTO kaomoji_catalog VALUES (?1,'match',?2)",
-            rusqlite::params![text, index],
+            rusqlite::params![text, index as i64],
         )
         .unwrap();
         db.execute(
             "INSERT INTO symbol_catalog VALUES (?1,'fixture','fixture','match',?2)",
-            rusqlite::params![text, index],
+            rusqlite::params![text, index as i64],
         )
         .unwrap();
     }
